Refactored PostResponse
the field 'likes' was a wrong type - It was written String[], but the API expects it to be UserResponse[].
the field 'comments' was a wrong type - It was written String[], but the API expects it to be CommentResponse[].
added a field previously missing - 'public'. It is named 'isPublic', because 'public' is a forbidden name for a variable. It has the annotation @SerializedName("public") which maps it as if it was 'public', so that the serialization/deserialization is successful.